+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 0.6.4
 * Added excludedParameters in CatcherOptions.
 * Updated dependencies
+* [BREAKING_CHANGE] Added isContextRequired parameter for ReportHandler.
+* [BREAKING_CHANGE] ReportHandler handle method requires additional BuildContext parameter.
+* [BREAKING_CHANGE] Toast handler supports now all platforms.
 
 ## 0.6.3
 * Updated sentry version.